+++ CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-## Chanded
+## Changed
 - `insert` returns ID of the object inserted
 - `multiInsert` returns an `array` with ID of the objects inserted
 - `save` returns ID of the object if it was inserted, or the Unique Index used as filter if it was updated
+- `multiSave` returns an `array` with ID of the objects inserted/updated
 
 ## [1.3.0] - 2019-07-30
 ### Added